Mudanças Tubarao: Chronometer para paradas, deletar teste, distancia percorrida salvando, calcula a dp estimada com a formula 1, splash screen provisória, deletar dentro do perfil, todas as formulas na classe calcula, analise teste abrindo no fim do teste.

+++ app/src/main/java/br/edu/ufcspa/tc6m/modelo/Teste.java
@@ -61,7 +61,7 @@ public class Teste implements Serializable {
 
     private String obsFinal;
 
-    private double distanciaPercorrida;
+    private int distanciaPercorrida;
     private double dpEstimada1;
     private double dpEstimada2;
     private double dpPorcento1;
@@ -208,11 +208,11 @@ public class Teste implements Serializable {
         this.estatura = estatura;
     }
 
-    public double getDistanciaPercorrida() {
+    public int getDistanciaPercorrida() {
         return distanciaPercorrida;
     }
 
-    public void setDistanciaPercorrida(double distanciaPercorrida) {
+    public void setDistanciaPercorrida(int distanciaPercorrida) {
         this.distanciaPercorrida = distanciaPercorrida;
     }
